requirements.txt
langchain-openai
openai
fastapi
uvicorn[standard]
gunicorn
python-dotenv
faiss-cpu
sentence-transformers 
psycopg2-binary
pgvector
requests

